requirements.txt
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
pyaes==1.6.1
pyasn1==0.6.1
python-dotenv==1.0.1
python-telegram-bot==21.8
requests==2.32.3
rsa==4.9
sniffio==1.3.1
Telethon==1.38.1
urllib3==2.2.3